add static file generator
